+++ ReboundPlugin.cs
@@ -18,12 +18,16 @@ namespace Rebound
     [BepInDependency("com.yuril.MovementPlus", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("ConfigTrixAirTricks", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("com.Dragsun.BunchOfEmotes", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("com.yuril.QuickTurn", BepInDependency.DependencyFlags.SoftDependency)]
     public class ReboundPlugin : BaseUnityPlugin
     {
         internal static ManualLogSource Log;
         public static Player player { get { return WorldHandler.instance?.GetCurrentPlayer(); }}
         internal static Harmony Harmony = new Harmony("goatgirl.Rebound");
+        
         public static bool MPlusActive = false; 
+        public static bool hasCBMono = false;
+        public static bool hasQuickTurn = false; 
 
         public static float landTime = float.PositiveInfinity;
         public static Vector2 landingVelocity = new Vector2(0, 0);
@@ -60,6 +64,13 @@ namespace Rebound
         public static bool rebounding = false;
         public static bool settingUpRebound = false;
 
+        public static Shader unlitShader { get { return Shader.Find("Unlit/Normal"); }}
+        public static bool hasUnlitCharacter = false; 
+
+        public static bool attemptingBoostedRebound { get { return player.CheckBoostTrick() && allowBoostedRebounds; } }
+        public static bool attemptingBurstRebound { get { return attemptingBoostedRebound && Vector3.Dot(player.dir, player.moveInput) <= -RBSettings.config_burstReboundSens.Value
+                                                                && player.moveInputPlain.sqrMagnitude >= (RBSettings.config_burstReboundDeadSquared); } }
+
         private void Awake()
         {
             ReboundPlugin.Log = base.Logger;
@@ -74,6 +85,16 @@ namespace Rebound
                     MPlusActive = true; 
                     Log.LogInfo($"Rebound: MovementPlus found!");
                 }
+                else if (plugin.Value.Metadata.GUID.Equals("com.yuril.QuickTurn"))
+                { 
+                    hasQuickTurn = true; 
+                    Log.LogInfo($"Rebound: QuickTurn found!");
+                }
+                else if (plugin.Value.Metadata.GUID.Equals("CrewBoom"))
+                { 
+                    hasCBMono = true; 
+                    Log.LogInfo($"Rebound: CrewBoomMono found!");
+                }
             }
 
             RBSettings.UpdateSettings(Config);
@@ -82,7 +103,6 @@ namespace Rebound
         private void FixedUpdate()
         {
             if (!InGame()) { return; }
-
             if (landTime < maxLandingTimeToRebound) { landTime += Reptile.Core.dt; } 
             
             if (!player.IsGrounded()) {
@@ -110,10 +130,31 @@ namespace Rebound
                 player.boostTrailTimer += Mathf.Min(Core.dt, player.GetVelocity().y/500f)*6f; 
                 player.trailWidth += Mathf.Min(Core.dt, player.GetVelocity().y/500f); 
             }
+
+            if (RBSettings.config_practiceMode.Value) {
+                if (CanReboundDetailed() && !hasUnlitCharacter) {
+                    hasUnlitCharacter = true;
+                    List<Material> playerMaterials = (hasCBMono ? RBCBHelper.GetPlayerMaterials(player)
+                                : new List<Material>() {player.characterVisual.GetComponentInChildren<SkinnedMeshRenderer>().material});
+                    foreach (Material material in playerMaterials) {
+                        material.shader = unlitShader; 
+                        if (material.HasColor("Main Color")) {
+                            material.SetColor("Main Color", new Color(255, 246, 143, 255)); // doesn't work due to unlit shader not being loaded 
+                        }
+                    }
+                } else if (!CanReboundDetailed() && hasUnlitCharacter) {
+                    hasUnlitCharacter = false; 
+                    player.SetOutfit(Reptile.Core.Instance.SaveManager.CurrentSaveSlot.GetCharacterProgress(player.character).outfit);
+                }
+            }
         }
 
         public static bool CanRebound() {
             return ReboundPlugin.landTime < maxLandingTimeToRebound && player.IsGrounded();
+        }
+
+        public static bool CanReboundDetailed() { // full requirements cross referenced from PlayerPatches
+            return CanRebound() && (!player.jumpConsumed || player.ability == player.slideAbility || player.ability == player.boostAbility);
         }
 
         public static bool PlayerAttemptingRebound() {
@@ -126,14 +167,15 @@ namespace Rebound
         }
 
         public static void ReboundTrick() {
-            bool isBoosting = allowBoostedRebounds ? player.boosting : false;
+            bool doingBoostedRebound = attemptingBoostedRebound;
+            bool doingBurstRebound = attemptingBurstRebound || (attemptingBoostedRebound && RBQTHelper.AbilityIsQuickTurn(player));
             float floorAngle = Vector3.Dot(Vector3.ProjectOnPlane(player.motor.groundNormalVisual, Vector3.up).normalized, player.dir);
             
             // Force (default) Rebound animations to properly transition to jump/fall animations
             player.animInfosSets[(int)player.moveStyle][Animator.StringToHash("jumpTrick1")].fadeTo[player.fallHash] = 1f;
             player.animInfosSets[(int)player.moveStyle][Animator.StringToHash("jumpTrick1")].fadeTo[Animator.StringToHash("fallIdle")] = 1f;
             
-            PrepForRebound(); // Set up variables, play SFX/voice, particle effects, etc.
+            PrepForRebound(doingBoostedRebound); // Set up variables, play SFX/voice, particle effects, etc.
             if (refreshCombo) { player.comboTimeOutTimer = landingComboMeter; }
             rebounding = true;
             settingUpRebound = true;
@@ -143,8 +185,8 @@ namespace Rebound
             reboundVelocity.y = -reboundVelocity.y*reboundVelocityMultiplier; 
             
             // Set trick animation and name
-            if (player.CheckBoostTrick() && allowBoostedRebounds) {
-                player.DoTrick(Player.TrickType.AIR, "Boosted Rebound", 0);
+            if (doingBoostedRebound) {
+                player.DoTrick(Player.TrickType.AIR, (doingBurstRebound ? "Burst Rebound" : "Boosted Rebound"), 0);
                 player.ActivateAbility(player.airTrickAbility);
                 reboundVelocity.x = Mathf.Max(reboundVelocity.x, player.GetForwardSpeed());
             } else { 
@@ -180,13 +222,26 @@ namespace Rebound
                 reboundVelocity.y = maxReboundYVelocity;
             }
 
+            // (NEW!) Boosted Rebound changes
+            if (doingBoostedRebound) {
+                reboundVelocity *= 1.1f; 
+                if (doingBurstRebound) {
+                    Vector2 oldReboundVelocity = reboundVelocity;
+                    reboundVelocity.y *= (1f - RBSettings.config_burstReboundPower.Value)*(1f + (RBSettings.config_burstReboundTransfer.Value));
+                    reboundVelocity.x = Mathf.Pow(reboundVelocity.x, RBSettings.config_burstReboundPower.Value); 
+                    reboundVelocity.y += (oldReboundVelocity.x - reboundVelocity.x)*RBSettings.config_burstReboundTransfer.Value;
+                    player.audioManager.PlaySfxGameplay(SfxCollectionID.GenericMovementSfx, AudioClipID.launcher_woosh,
+                    player.playerOneShotAudioSource, 0f);
+                }
+            }
+            
             // Adjust velocity based on floor normal
             Vector2 floorAngleXY = new Vector2(
                 Mathf.Sin(floorAngle)*(RBSettings.config_slopeMultiplierX.Value), 
                 Mathf.Cos(floorAngle)*(RBSettings.config_slopeMultiplierY.Value));
             if (floorAngleXY.x < 0f) { floorAngleXY.y += 1f; } // higher instead of lower
-            //floorAngleXY.y = Mathf.Lerp(floorAngleXY.y, 1f, 0.5f);
 
+            // Apply Rebound velocity
             player.motor.SetVelocityYOneTime(reboundVelocity.y * floorAngleXY.y);
             float newForwardSpeed = reboundVelocity.x + (reboundVelocity.y * floorAngleXY.x);
             player.SetForwardSpeed(newForwardSpeed);
@@ -216,8 +271,7 @@ namespace Rebound
             settingUpRebound = false;
         }
 
-        public static void PrepForRebound() {
-            bool isBoosting = allowBoostedRebounds ? player.boosting : false;
+        public static void PrepForRebound(bool isBoosting) {
             if (!isBoosting) { player.StopCurrentAbility(); }
             
             player.jumpedThisFrame = true;